// src/app/models/depo.ts
export class Depo {
  srk_no: number;
  dp_no: number;
  dp_ad: string = "";
  dp_stok_tur: number = 0;
  uk: string = "";
  updt: any | null = null;
  iuk: string = "";
  idt: any | null = null;
  dp_c1: string = "";
  dp_c2: string = "";
  dp_ent_kod: string = "";
  dp_stok_kont_sw: number = 0;
}