// src/app/models/vardiya.ts
export class Vardiya {
  srk_no: number;
  vad_kod: string;
  vad_ad: string = "";
  vad_bas: Date | null = null;
  vad_bit: Date | null = null;
  vad_gunsonu: Date | null = null;
  vad_gunsonu_isrt: number = 0;
  uk: string = "";
  updt: Date | null = null;
  iuk: string = "";
  idt: Date | null = null;
  vad_sira: number;
}