// src/app/models/stkfdtop.ts
export class Stkfdtop {
  sfd_primno: number = 0;
  sfd_sfd_primno: number = 0;
  sfd_sf_primno: number = 0;
  sfd_srk_no: number = 0;
  sfd_bcmno: number = 0;
  sfd_yil: number = 0;
  sfd_seri: string = '';
  sfd_dp_no: number = 0;
  sfd_no1: number = 0;
  sfd_no2: number = 0;
  sfd_sira: number = 0;
  sfd_tpsira: number = 0;
  sfd_frm_kod: string = '';
  sfd_mm_primno: number = 0;
  sfd_mm_tur: number = 0;
  sfd_mm_kod: string = '';
  sfd_trh: Date | null = null;
  sfd_ay: number = 0;
  sfd_fist_no: number = 0;
  sfd_mkt: number = 0;
  sfd_birim: string = '';
  sfd_mkt_mt: number = 0;
  sfd_mkt_top: number = 0;
  sfd_brtmkt: number = 0;
  sfd_brtmkt_mt: number = 0;
  sfd_brtmkt_top: number = 0;
  sfd_mkt_amb: number = 0;
  sfd_grp_partino: string = '';
  sfd_partino: string = '';
  sfd_lotno: string = '';
  sfd_harmanno: string = '';
  sfd_frm_sipno: string = '';
  sfd_mark_kod: string = '';
  sfd_aciklama: string = '';
  sfd_en: number = 0;
  sfd_grmj: number = 0;
  sfd_pus: number = 0;
  sfd_fayn: number = 0;
  sfd_metretul: number = 0;
  sfd_tup_may: string = '';
  sfd_sput_primno: number = 0;
  sfd_spu_primno: number = 0;
  sfd_sw_primno: number = 0;
  sfd_sw_bcmno: number = 0;
  sfd_sw_islno: number = 0;
  sfd_spd_primno: number = 0;
  sfd_spd_sipno: number = 0;
  sfd_sde_no: number = 0;
  sfd_spd_no1: number = 0;
  sfd_spd_no2: number = 0;
  sfd_spd_sira: number = 0;
  sfd_sw_wno: number = 0;
  sfd_spu_sira: number = 0;
  sfd_topsira: number = 0;
  sfd_topsira2: number = 0;
  sfd_stk_yer1: string = '';
  sfd_yer_mkt1: number = 0;
  sfd_yer_top1: number = 0;
  sfd_yer_bmkt1: number = 0;
  sfd_fiyat: number = 0;
  sfd_tutar: number = 0;
  sfd_dvz_kod: string = '';
  sfd_iskonto: number = 0;
  sfd_vade: number = 0;
  sfd_hspbrm: string = '';
  sfd_hsptur: string = '';
  sfd_tutar_gir: number = 0;
  sfd_tutar_cik: number = 0;
  sfd_rspd_primno: number = 0;
  sfd_rspd_no1: number = 0;
  sfd_rspd_sipno: number = 0;
  sfd_mak_kod: string = '';
  sfd_tms_kod: string = '';
  sfd_gsb_kod: string = '';
  sfd_irs_no: number = 0;
  sfd_irs_trh: Date | null = null;
  sfd_irs_frm_kod: string = '';
  sfd_tlmt_primno: number = 0;
  sfd_tlmt_kod: string = '';
  sfd_cl_primno: number = 0;
  sfd_cl_kod: string = '';
  sfd_amblj_birim: string = '';
  sfd_ksl_primno: number = 0;
  sfd_ksl_no1: number = 0;
  sfd_ksl_no2: number = 0;
  sfd_ksl_sira: number = 0;
  sfd_ksl_tpsira: number = 0;
  sfd_barkod: string = '';
  sfd_klt_drm: number = 0;
  sfd_kurtrh: Date | null = null;
  sfd_kur: number = 0;
  uk: string = '';
  updt: Date | null = null;
  iuk: string = '';
  idt: Date | null = null;
  sfd_belgeno: number = 0;
  sfd_belge_trh: Date | null = null;
  sfd_spd_primno_wo: number = 0;
  sfd_spd_bcmno_wo: number = 0;
  sfd_spd_no1_wo: number = 0;
  sfd_spd_no2_wo: number = 0;
  sfd_spd_sira_wo: number = 0;
  sfd_vardiya: string = '';
  sfd_per_no: number = 0;
  sfd_cld_primno: number = 0;
  sfd_cld_sira: number = 0;
  sfd_trh_cik: Date | null = null;
  sfd_ay_cik: number = 0;
  sfd_fist_no_cik: number = 0;
  sfd_irs_no_cik: number = 0;
  sfd_irs_trh_cik: Date | null = null;
  sfd_irs_frm_kod_cik: string = '';
  sfd_barkod2: string = '';
  sfd_kdv_oran: number = 0;
  sfd_kdv_tutar: number = 0;
  sfd_isk_tutar: number = 0;
  sfd_parti_isrt: number = 0;
  sfd_prs_primno: number = 0;
  sfd_prs_kod: string = '';
  sfd_dara1: number = 0;
  sfd_dara2: number = 0;
  sfd_hspdvz_kod: string = '';
  sfd_fsw_bcmno: number = 0;
  sfd_fspd_primno: number = 0;
  sfd_fsw_primno: number = 0;
  sfd_fspd_no1: number = 0;
  sfd_fspd_no2: number = 0;
  sfd_fspd_sipno: number = 0;
  sfd_fspd_sira: number = 0;
  sfd_fsw_wno: number = 0;
  sfd_saspd_primno: number = 0;
  sfd_saspd_no1: number = 0;
  sfd_saspd_no2: number = 0;
  sfd_saspd_sipno: number = 0;
  sfd_saspd_sira: number = 0;
  sfd_orgu_frm_kod: string = '';
  sfd_total_isk_tutar: number = 0;
  sfd_brut_tutar: number = 0;
  sfd_sp_frm_kod: string = '';
  sfd_dvz_fiyat: number = 0;
  sfd_kdv_tutar_tev: number = 0;
  sfd_fyt_tl: number = 0;
  sfd_des_id: number = 0;
  sfd_des_kod: string = '';
  sfd_varyant: string = '';
  sfd_ack2: string = '';
  sfd_ip_uzun: string = '';
  sfd_tutar_tl: number = 0;
  sfd_mhsp_mkod: string = '';
  sfd_mhsp_mid: number = 0;
  sfd_kdvhsp_mid: number = 0;
  sfd_rzv_id: number = 0;
  sfd_sprt_id: number = 0;
  sfd_eirs_no: string = '';
  //sfd_lme: number = 0;
  sfd_prim: number = 0;
  StkfCmpt?: StkfCmpt;

  Where?: any;
}

export class StkfCmpt {
  cmpt_mm_ad: string = '';
  cmpt_sf_isl_tutar: number = 0;
  cmpt_previous_sf_primno: number = 0;
  cmpt_sf_primno: number = 0;
}